requirements.txt
ultralytics
opencv-python
numpy
torch
omegaconf
streamlit
pillow